Create client_mobile.py
Satisfy the demand of running client on mobile devices.
Make server ip and username input easier.
Cancel the timer based on tkinter module, which module is not widely  supported by interpreters on mobile devices.
The timer in this program is based on the terminal, which reminds you of your remaining time with a 5 seconds' interval.